job-bot/requirements.txt
playwright
schedule
pymongo
requests
openai
python-dotenv
